Added path finding options Implemented Structs as Arrays for solver.rs Added ability for multiple solvers at once.

+++ src/types.rs
@@ -1,4 +1,4 @@
 pub type Cost = f32;
 pub type Flag = u32;
 pub type Pos = f32;
-pub type Index = usize;
+pub type Index = u32;
